fastapi-app/requirements.txt
fastapi==0.85.0
uvicorn==0.18.3
transformers==4.21.2
torch==2.4.1
pydantic==2.0.2
python-dotenv==0.21.0